refactor(panelStore): split panelStoreListeners into per-domain modules
- Extract pure identity reducers (reduceAgentDetected, reduceAgentExited)
  into src/store/listeners/panel/identityReducer.ts with unit tests
- Split listener subscriptions into identity/lifecycle/activity/
  backendHealth/resource sub-modules under src/store/listeners/panel/
- Each sub-module returns its own DisposableStore; coordinator composes
  via masterStore.add(subStore) so children dispose in LIFO
- Move recoveryTimer, fallbackInFlight, activityBuffer to their owning
  sub-modules with explicit cleanup disposables
- Gate _identityLog, recordIdentityEvent, window.__daintreeIdentity*
  globals, and all [IdentityDebug] console.log calls behind
  import.meta.env.DEV so Rolldown DCE eliminates them in production
- Coordinator keeps the single import.meta.hot.dispose+accept boundary
  (sub-modules deliberately have no HMR blocks) and re-exports the
  stable public API (setupTerminalStoreListeners, cleanupTerminalStoreListeners)

Closes #5922
